Create POC of working with audio
Create 3 POCs:
- One POC of playing the sound as it is recorded
- One POC of recording audio to file (using pickle)
- One POC of playing audio from file (using pickle)

Each of the POCs work, but none of them work in the same way we want the
final components to work. I do not know how robust each of these POC is,
and if they will work in any computer without altering them (for example
if the default device is not the one you want, it will not work).

I still don't know how to specify in the recording only POC that I want
the data to go into a numpy array.

We still need to try a POC of passing the audio over TCP (and decide how
to pass it - maybe just struct.pack?)