Add non blocking option to visualizer
In order to support the visualize from net mode, I needed to add a non blocking mode to the visualizer.
In non blocking mode, you can decide how long you want to let the visualizer block.

This is needed, as the network receiver needs run time in the main thread in order to receive data,
unlike the recorder receiver that uses another thread in order to record in the background.