Make everything close nicely when socket sender runs out of time
Change both the socket sender and socket receiver to close without
exceptions when the socket sender runs out of time and closes its
socket.

We probably won't use this feature so much, but it's good practice
to not throw an exception when a socket closes and instead close
everything nicely and return gracefully.