Add option to make the wait function of SocketWriter not block
Change the timeout argument of SocketWriter, so that when you put a
value that is not None in it, the wait funciton will only block for this
long and then will return to the reader.

Added a complex example that uses sockets to the yamls for the run script,
and updated the readme with instuctions about how to run it.